+++ ebegu-server/src/main/java/ch/dvbern/ebegu/rules/FachstelleAbschnittRule.java
@@ -50,7 +50,7 @@ public class FachstelleAbschnittRule extends AbstractAbschnittRule {
 	@Nonnull
 	private VerfuegungZeitabschnitt toVerfuegungZeitabschnitt(@Nonnull PensumFachstelle pensumFachstelle) {
 		VerfuegungZeitabschnitt zeitabschnitt = new VerfuegungZeitabschnitt(pensumFachstelle.getGueltigkeit());
-		zeitabschnitt.setFachstellenpensum(pensumFachstelle.getPensum());
+		zeitabschnitt.setFachstellenpensum(pensumFachstelle.getPensum().intValue());
 		return zeitabschnitt;
 	}
 }
