+++ ebegu-shared/src/main/java/ch/dvbern/ebegu/entities/AbstractPensumEntity.java
@@ -40,17 +40,17 @@ public class AbstractPensumEntity extends AbstractDateRangedEntity {
 	@Min(0)
 	@NotNull
 	@Column(nullable = false)
-	private Integer pensum;
+	private Long pensum;
 
 	public AbstractPensumEntity() {
 	}
 
 	@Nonnull
-	public Integer getPensum() {
+	public Long getPensum() {
 		return pensum;
 	}
 
-	public void setPensum(@Nonnull Integer pensum) {
+	public void setPensum(@Nonnull Long pensum) {
 		this.pensum = pensum;
 	}
